requirements.txt
# local package
# -e .

# external requirements
click
python-dotenv>=0.5.1
loguru

# data
# awscli

streamlit
pandas
scikit-learn
fastapi
uvicorn
pygwalker
pretty_errors
